requirements.txt
bertopic
pandas
numpy
scikit-learn
sentence-transformers
umap-learn
hdbscan
tqdm
streamlit
groq
scipy
json5
plotly
nltk
requests